tutorials/python/EX1_dessign_patterns/pyproject.toml
[tool.poetry]
name = "dp"
version = "0.1.0"
description = "Design patterns for Python ML Projects"
authors = ["Juan Gonzalez <jucgonzalezes@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2"
google-cloud-bigquery = "^3.25.0"
google-cloud-bigquery-storage = "^2.26"  # To speedup BQ calls
db-dtypes = "^1.3"  # To interface between pandas and BQ
polars = "^1.6.0"
scikit-learn = "^1.5.1"
catboost = "^1.2.7"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.8.0"
pytest = "^8.3.2"
mypy = "^1.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
